+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+<a name="1.0.6"></a>
+## [1.0.6](https://github.com/poppinss/edge-lexer/compare/v1.0.4...v1.0.6) (2018-07-10)
+
+
+### Code Refactoring
+
+* **tokenizer:** make use of edge-error for consitent errors ([f896e77](https://github.com/poppinss/edge-lexer/commit/f896e77))
+
+
+### BREAKING CHANGES
+
+* **tokenizer:** Tokenizer needs an options object with filename
+
+
+
 <a name="1.0.5"></a>
 ## [1.0.5](https://github.com/poppinss/edge-lexer/compare/v1.0.4...v1.0.5) (2018-07-04)